requirements.txt
Flask==3.0.3
gunicorn==20.1.0
beautifulsoup4==4.12.3
requests==2.32.3
psycopg2-binary==2.9.9
azure-ai-translation-text==1.0.1
python-dotenv==1.0.0